move test infra to mock package
